requirements.txt
asgiref==3.7.2
dj-rest-auth==5.0.1
Django==4.2.5
django-cleanup==8.0.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3